+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+### Added
+
+* `add` parameter to `wn.download()` ([#73])
+* `--no-add` option to `wn download` command ([#73])
+* `progress_handler` parameter to `wn.download()` ([#70])
+* `progress_handler` parameter to `wn.add()` ([#70])
+
 ### Changed
 
 * Renamed `lgcode` parameter to `lang` throughout ([#66])
@@ -60,3 +67,5 @@ abandoned, but this is an entirely new codebase.
 [#60]: https://github.com/goodmami/wn/issues/60
 [#61]: https://github.com/goodmami/wn/issues/61
 [#66]: https://github.com/goodmami/wn/issues/66
+[#70]: https://github.com/goodmami/wn/issues/70
+[#73]: https://github.com/goodmami/wn/issues/73